Fixed basic requests handling code.

+++ util.c
@@ -34,7 +34,7 @@ log_printf(int level, char *fmt, ...)
 }
 
 void
-log_print_data(int level, size_t bytes, unsigned char *buf)
+log_print_data(int level, unsigned int bytes, unsigned char *buf)
 {
   if (level <= loglevel) {
     for (; bytes > 0; bytes--, buf++)
